shared/shared/requirements.txt
# FastAPI and dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.6

# Authentication
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6

# Networking
httpx>=0.24.1

# Storage
minio>=7.1.15

# Redis and caching
redis>=4.6.0
hiredis>=2.2.3

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.2